Move lazy_property to util package.
